+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- missing Accept header in GET requests
+
+## [5.2.0] - 2023-03-09
+### Added
+- `fetch` options can now take `baseUrl` for defining custom base URL for the single request only
+
+## [5.1.0] - 2023-03-08
+### Added
+- `getAccessToken` method for fetching the access token that's up-to-date
 
 ## [5.0.0] - 2022-10-05
 ### Changed
@@ -206,7 +216,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `verifyPasswordlessLogin` method to `BrainClient` which checks status of passwordless login
 - `getGithubLoginUrl` method to `BrainClient`
 
-[Unreleased]: https://github.com/superfaceai/service-client/compare/v5.0.0...HEAD
+[Unreleased]: https://github.com/superfaceai/service-client/compare/v5.2.0...HEAD
+[5.2.0]: https://github.com/superfaceai/service-client/compare/v5.1.0...v5.2.0
+[5.1.0]: https://github.com/superfaceai/service-client/compare/v5.0.0...v5.1.0
 [5.0.0]: https://github.com/superfaceai/service-client/compare/v4.0.0...v5.0.0
 [4.0.0]: https://github.com/superfaceai/service-client/compare/v3.0.0...v4.0.0
 [3.0.0]: https://github.com/superfaceai/service-client/compare/v2.1.0...v3.0.0